requirements.txt
pyserial==3.5
numpy==1.19.2